backend/requirements.txt
# OmniAI Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
greenlet==3.0.3

# Redis
redis==5.0.1
aioredis==2.0.1

# Auth
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
httpx==0.26.0

# File Processing
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0

# Web Search
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
python-magic==0.4.27